Move config db testing methods to utility class
Move test utility classes for DbResourceGroupConfigurationManager to H2TestUtil
